refactor: 입력 관련 예외처리 파일 분리
